Added Creditor class. Main uses the CreditorManager which uses Creditor. Working on creditor input loop.

+++ src/DraftCalculator.java
@@ -1,6 +1,6 @@
 public class DraftCalculator {
 
-    public int provideEstimate(int totalDebt, double FEE_RATE) {
+    public double provideEstimate(int totalDebt, double FEE_RATE) {
 
         //Setting estimate to 60%
         return (totalDebt * 0.6) + (totalDebt+FEE_RATE);
